render_backend_youtube_transcriber/requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
ffmpeg-python